Move user theme selection to user store
