+++ client/src/composables/user.ts
@@ -11,24 +11,31 @@ import type { Store } from "vuex";
 export function useCurrentUser(noFetch: boolean | Ref<boolean> = false) {
     // TODO: add store typing
     const store = inject("store") as Store<unknown>;
-
     const currentUser = computed(() => store.getters["user/currentUser"]);
-
     onMounted(() => {
         if (!unref(noFetch)) {
             store.dispatch("user/loadUser");
         }
     });
-
     const addFavoriteTool = async (toolId: string) => {
         await store.dispatch("user/addFavoriteTool", toolId);
     };
-
     const removeFavoriteTool = async (toolId: string) => {
         await store.dispatch("user/removeFavoriteTool", toolId);
     };
-
     return { currentUser, addFavoriteTool, removeFavoriteTool };
+}
+
+export function useCurrentTheme() {
+    const store = inject("store") as Store<unknown>;
+    const currentTheme = computed(() => store.getters["user/currentTheme"]);
+    function setCurrentTheme(theme: string) {
+        store.dispatch("user/setCurrentTheme", theme);
+    }
+    return {
+        currentTheme,
+        setCurrentTheme,
+    };
 }
 
 // temporarily stores tags which have not yet been fetched from the backend
@@ -39,24 +46,18 @@ const localTags = ref<string[]>([]);
  */
 export function useUserTags() {
     const { currentUser } = useCurrentUser(true);
-
     const userTags = computed(() => {
         let tags: string[];
-
         if (currentUser.value) {
             tags = [...currentUser.value.tags_used, ...localTags.value];
         } else {
             tags = localTags.value;
         }
-
         const tagSet = new Set(tags);
-
         return Array.from(tagSet).map((tag) => tag.replace(/^name:/, "#"));
     });
-
     const addLocalTag = (tag: string) => {
         localTags.value.push(tag);
     };
-
     return { userTags, addLocalTag };
 }
