+++ client/src/store/userStore/queries.js
@@ -35,17 +35,7 @@ export async function setCurrentTheme(userId, themeId) {
     const url = prependPath(`/api/users/${userId}/theme/${themeId}`);
     try {
         const { data } = await axios.put(url, { theme_id: themeId });
-        return data["theme_id"];
-    } catch (e) {
-        rethrowSimple(e);
-    }
-}
-
-export async function getCurrentTheme(userId) {
-    const url = prependPath(`/api/users/${userId}/theme`);
-    try {
-        const { data } = await axios.get(url);
-        return data["tools"];
+        return data;
     } catch (e) {
         rethrowSimple(e);
     }
